api: initial working C library
Issue: US100616
Issue: US100606

This commit contains the initial working source code for the new IoT
library.  It is a port of code from the old library/service model to
work ith the new cloud provider.  The following features are currently
working:
  - telemetry
  - location
  - actions (preliminary)

Additionally, features also included from the old code base are:
  - unit test support (with some unit tests copied over and compiling)
  - code coverage report generation
  - sample application and build system
  - configuration file reading and device id generation
  - exporting of operating system abstraction layer to an external
    repository

Currently, support is only work for Linux but preliminary work for
Android operating system is included in this repository.

Signed-off-by: Keith Holman <keith.holman@windriver.com>