build: fixes to build on Ubuntu
On Ubuntu 64-bit libraries are installed into the "/usr/lib" directory,
whereas on Fedora 64-bit libraries are installed into the "/usr/lib64"
directory.  This patch contains a fix that determines the correct
location based on CMake properties that store this information.
Additionally, this patch removes a find_package reference that is no
longer relavent.

Signed-off-by: Keith Holman <keith.holman@windriver.com>