build: fixes to avoid operating system pollution
This patch provides fixes when building the 3rd-party libraries to avoid
operating system pollution.  Some Linux operating systems were install
libraries to the directory "lib/[subdir]", this was causing the find
package module to fail.  This fix updates the build system to handle
this case.  Additionally, this patch contains fixes to ensure that the
update builds with OpenSSL libraries.

Signed-off-by: Keith Holman <keith.holman@windriver.com>