api: add unit tests for json encoder
This patch adds some unit testing for the JSON encoder.  The JSON
encoder API is considered deliverable for customers to use.  To ensure
that it will behave as expected unit testing of the interface was
introduced.  This patch fixes any issues found, as well as adds
commenting to the interface to instruct how to use the encoder.  The
testing passed with both the internal hand encoder as well as using
jansson under the hood.

Signed-off-by: Keith Holman <keith.holman@windriver.com>