api: add osal include directory to search path
If libuuid-dev is installed then "uuid/uuid.h" will be found in /usr/include.
This header file is included in the <os.h> header file, even though the
file it is trying to use is under it's own repository.   This fix
includes the OSAL layer include directory as a search path so that the
correct header file is found and used.

Signed-off-by: Keith Holman <keith.holman@windriver.com>