control: Initial implementation of iot-control
This is the initial port of the iot-control application from the HDC 2.x
repository.  This allows for the iot-control application to generate a
configuration file for connecting to the cloud.  This also contains a
basic initial implmentation of a JSON schema parser which is used by the
control application to generate and verify the configuration file.

Signed-off-by: Keith Holman <keith.holman@windriver.com>