json: add ability to cancel & clear an object
While encoding it is desirable sometimes to not encode an object if
during the logic it is determined that there are no items to add.  This
patch adds the functions iot_json_encode_object_cancel and
iot_json_encode_object_clear to help with these purposes.  Unit tested
with both jsmn (self encoder) and using the jansson encoder system.
Unit test was also valgrinded to remove all errors and memory leaks with
the encoder.  This patch also includes fixes in the declaration of
functions in the header file iot_json.h.

Signed-off-by: Keith Holman <keith.holman@windriver.com>