tr50.c: add support for retry and resume for file transfer.
This commit adds support for file upload/download retry and
resume.  There is a maximum number of retries setting in iot_defs.h,
where -1 is unlimited, and >0 is finite.

Note: the cloud has a file handle timeout of 300 seconds.  If the
download is interrupted for up to 300 seconds the download will fail.

Signed-off-by: Paul Barrette <paulbarrette@gmail.com>