device-manager: add "quit" method
This patch adds a "quit" method to closer align to the functionality of
the python agent.  Additionally, this patch includes a minor fix if an
item was not found in the "actions_enabled" configuration portion of a
config file it would go to "false", even if the compile default is
"true".

Signed-off-by: Keith Holman <keith.holman@windriver.com>