test: update base unit tests
In the API there was a bunch of unit tests that were commented out and
newer parts of the code base wasn't being unit tested.  This commit
updates the unit tests for the iot_base.c api file, by uncommenting and
updating existing unit tests and adding unit tests for missing
functionality.

Signed-off-by: Keith Holman <keith.holman@windriver.com>