test: use common initialize/finalize functions
This patch updates the test_support library used by the unit tests to
include an test_initialize and test_finalize functions and updates the
current unit tests to use the function.  This allows the ability to seed
all the unit tests to allow later exact reproducibility of a test
containing random numbers.

Signed-off-by: Keith Holman <keith.holman@windriver.com>