build: add option to ensure that a value is set
This patch add a function to cmake to ensure that an option will always
have a value set even if give a blank -D value on the command line.
For example: "cmake -DIOT_THREAD_SUPPORT:BOOL= ." was setting the value
IOT_THREAD_SUPPORT to a value of "" which would make it not be set to
the default value of "ON".  This fix ensures that this variable will be
either OFF, ON, or an explicitly set value on the command line.

Signed-off-by: Keith Holman <keith.holman@windriver.com>