tr50: fix mqtt messages being sent out
This patch updates the code surrounding the sending of mqtt messages by
the plug-in.  A common function was created to ensure that messages are
sent and logged in a consistent method.  This code also updates the
messages to ensure that all mqtt messages contain an id number, so that
in the future responses can tracked to out-going messages.  This patch
also cleaned up logging by removing log message that were prepended with
"-->" to now say either "sent" or "received" depending on whether the
device sent or received an mqtt message respectfully from the cloud.

Signed-off-by: Keith Holman <keith.holman@windriver.com>