build: use threads throughout code on Android
The flag `IOT_THREAD_SUPPORT=1` was being set at the base level, but not
being set throughout the product.  This leads to some parts of the
product being compiled with thread support and other parts without.
This patch changes the define from `LOCAL_CFLAGS` to `APP_CFLAGS`, so
that the value isn't cleared with a call to `include $(CLEAR_VARS)`.

Signed-off-by: Keith Holman <keith.holman@windriver.com>