build: fix use of PREFIX variable
This patch updates the `generate_header.sh` script to make correct use
of the `PREFIX` variable.  This variable is to be used in the future to
change the library prefix if desired for IOT to something else.
However, if this is changed then the script wouldn't work as intended.
This patch is a simple fix to allow the variable to work in the script
now although, other parts of the build system would have to be changed
to support a change of the library PREFIX.

Signed-off-by: Keith Holman <keith.holman@windriver.com>