tr50: merge code for connect & reconnect
This patch updates the code-base for the tr50 plug-in to use the same
code for connecting & reconnecting.  For example, if connecting over a
proxy server using mosquitto the code would not reconnect using a proxy
server because initial connection code was copied but proxy support was
only added to the connect (and not reconnect) method.  This refactoring
ensures the same settings are used for both connecting and
reconnecting.

Signed-off-by: Keith Holman <keith.holman@windriver.com>