api: allow connecting when network not available
In the case where the network is not available the function
`iot_mqtt_connect` would not work.  This fix allows the function to loop
until the time out is reached or a connection is established.  To test
this case start up an application with the outside network interface
down. `ifconfig eno1 down`, then after starting the application, bring
the network interface up `ifconfig eno1 up`.  The application should
eventually connect without intervention. This loop respects the timeout
passed to it.  So passing a timeout will break the loop if a connection
cannot be fully established.

Signed-off-by: Keith Holman <keith.holman@windriver.com>