api: show receive/send messages as debug
Setting the log level to "TRACE", causes a lot of output as expected
that isn't totally helpful to non-developers.  However, setting it up a
level to "DEBUG", causes not enough information to see if there are
messages going in and out of the system.  This patch updates the code to
show raw messages that come in and out on the DEBUG level, so that there
is a good level of intermediate of information being produced.  If this
isn't desired than a user can set the level higher to "INFO".

Signed-off-by: Keith Holman <keith.holman@windriver.com>