tr50: always use 3 digits for milliseconds
fixes: #103

When publishing a timestamp in a tr50 message, the milliseconds where
just being added as the number of milliseconds.  However, this was
misleading as on the cloud, as 1 or 2 digit milliseconds where missing
leading 0's.  For example, .57 was being represented instead of the
intended .057.  This causes the cloud to not show the events in the
correct order.

Signed-off-by: Keith Holman <keith.holman@windriver.com>