api: handle short device id's
fixes: 113

Previously there was an error reading the device_id file containing an
id < 36 characters and containing a "new-line" character.   The MQTT
connections was using the new-line character as part of the client id,
causing the connection to not be accepted by the cloud.  This patch
fixes the reading to test if there are any valid ascii characters in the
device id file.

Signed-off-by: Keith Holman <keith.holman@windriver.com>